getPair must be total
getPair used to be designed to be a partial function, meaning it does not always have a result and the wallet-common had to do the fallback. instead, with this change, the responsability to fallback will be on user-land. it simplifies code of this lib but also allows more freedom to fallback as we want.
atm the mobile code is ready for it, but the desktop will need a slight change to not have this being a regression