bugfix fetchHistodayRates() to return the close time
previously it was not correctly returning the close date for the last day.
the last day was actually to day of today, but today shouldn't be returned because it's not yet finished,
therefore the "close" value was always different if you would have asked fetchHistodayRates() multiple times.
we now only consider historic days that are finished and "stable" and at the "close time" (23:59:59)
if you ask the countervalue of today, it should fallback on "latest" because the day is not yet closed