[cosmos] Change broadcast mode from block to sync
This change has been tested on LedgerLiveBot and doesn't trigger any
issues regarding synchronizing and optimistic updates / finding the
operations.

This change, as stated by the comment above the patch, moves the
broadcast call to "intermediate" level of blocking. It basically returns
just before the transaction enters the mempool, but the backend node
will still verify the transaction in a stateless manner and send an
error if anything wrong happens.

It is done to mitigate the "RPC timeout issues" we might encounter