Token support and further swap development (#786)
* SwapHistory reconciliation/export/import for liveqr

* USDT signature for swap cli testing

* Adapt the logic to also work with tokens.

Added two signatures/configs for erc20 token testing, currently the usdt does NOT work due to a ticker collision but augur does work. We can pass the tokenId (as in full erc20.js id) on the command signature for cli

* Adapt Swap History to erc20 tokens without funds yet

Allow swap operations for tokens that dont have funds yet to still show on the list as if we had an account for them. This is done by enhancing the main account and checking if the token account already existed in the subaccount list. We need this in order to prevent navigation to an account that doesn't really exist in our data.

* Make swap history not optional in accounts

* Use hex strings for swap sigs and lint

* lint

* regen snapshot

* PR Review changes

* Add missing pending state for swap status

Co-authored-by: Gaëtan Renaudeau <renaudeau.gaetan@gmail.com>