Fix token APY from Compound API (#990)
* Fix token APY from Compound API

I'm no expert in Compound, so maybe this comes from a confusion on my part. Don't hesitate to tell me if I'm wrong.

I think what is currently fetch (`comp_supply_APY`) is the APY in COMP tokens when supplying an asset. What should be display is (at least) the APY in the supplied token itself, isn't it?

In the Compound `/ctoken` API, the right field to use seems to be `supply_rate`, which is in decimal form (we need to multiply it by 100 to have percentage).

* prettier

Co-authored-by: Gaëtan Renaudeau <renaudeau.gaetan@gmail.com>