Fix the reported Compound supply values – LL-5470 (#1164)
* Fix the reported Compound supply values

Previously, the reported supply balance was inaccurate. First, we were
using the transaction time to look up the historical exchange rate. If
the tx was in the mempool for a while, this could create a significant
delta to when the block was written.

* Updated fetchHistoricalRates to use the block depth, this is more
  accurate than querying based on transaction time.
* Use the underlying token transaction instead of using the historical
  rate for calculation. Even the block number-based implementation has
  noticeable error.

Technically, fetchHistoricalRates isn't needed anymore, but I have left it in
place to not affect anything that may depend on `extra.rate`.

* prettier

* fix flow type errors

* reduce number of loops

Co-authored-by: JunichiSugiura <jun.sugiura.jp@gmail.com>
Co-authored-by: Gaëtan Renaudeau <renaudeau.gaetan@gmail.com>