Allow connectApp command to have dependencies (#1156)
* Add optional dependencies to connectApp

In order to allow flows to have dependent application installed we introduce the dependencies array which will check (and attempt to install) the installed apps on the device before attempting to open the target one.

* Make connectApp changes non breaking

Keep the event name and signature compatible with current implementation so that a transition towards multiple missing apps is not a breaking change.

* Autoquit if we have dependencies and we're not in dashboard

In order to perform the listapps that happens inside the streamAppInstall code, we need to be in the dashboard. There was an edge case where we have dependencies and the user is already in the target app, we would be skipping the dependency check. Nice catch @gre

* refactor state for stream install apps

Co-authored-by: Gaëtan Renaudeau <renaudeau.gaetan@gmail.com>