atom(button): allowing colors override
There are some special case when we need a unique variant not managed by
the theme. In that case, we should be able to override the color, the
background-color and the border-color css properties.
I move the default values of these properties at the beginning of the
stylesheet to allow cascading override. In addition to that, I inject
the border system to the base button component.