tweak PR template to make checklist good to check
If you check all the points of a checklist, it means you are "good". By making "Breaking changes" and not "No breaking changes" a checklist point, it confuses a bit that fact, and it's actually not good to check it. When you don't check a check list, it's where you need to explain why. I shouldn't explain why i'm not breaking =)