feat(Firmware Update Device SDK): create a device action in the SDK for firmware updates
This is the complete Device Action for firmware updates in the new Device SDK it includes
all the commands tasks and the action state reducer so it can be used in any JS based app.

It also includes an entry point in our CLI for testing the device action and a hook to use it
in any React App.

A few modifications in the general workings of the SDK had to be made such as the creation of
a retry mechanism for handling device disconnections when it quits an app or when the device reboots
during an update

Co-authored-by: Alexandre Magaud <alexandre.magaud@gmail.com>