chore: refactoring transport-node-hid-singleton
Strategy: disconnect after a timeout once close is called.
And clear disconnection timeout when open is called.
No more external preventing/not-preventing auto-disconnect

Also removed unnecessary setEnableDisconnectAfterInactivityForTransport.
It was meant for the TransportNodeHid. But it was called inside withDevice, which
is called inside the renderer process. So this information was not propagated
to the internal process. And consequently is unnecessary anymore.

Also fix: Promise reject and not error throw